refactor: main 내 폴더 분리
