+++ apps/front/src/components/Main/components/Tab/index.module.scss
