+++ apps/front/src/components/Main/data/index.ts
@@ -1,0 +1,7 @@
+export const CHAT = "chat";
+export const FRIEND = "friend";
+
+export const tabs = [
+  { id: CHAT, label: "채팅방" },
+  { id: FRIEND, label: "친구" },
+];
