+++ apps/front/src/components/Main/index.tsx
@@ -2,16 +2,11 @@ import { useState } from "react";
 import Chat from "components/Chat";
 import DefaultHeader from "components/DefaultHeader";
 import Friends from "components/Friends";
+import Tab from "./components/Tab";
+import { CHAT, FRIEND, tabs } from "./data";
 import styles from "./index.module.scss";
-import Tab from "./Tab";
 
 function Main() {
-  const CHAT = "chat";
-  const FRIEND = "friend";
-  const tabs = [
-    { id: CHAT, label: "채팅방" },
-    { id: FRIEND, label: "친구" },
-  ];
   const [currentTab, setCurrentTab] = useState<string>(CHAT);
   const handleTab = (e: React.MouseEvent<HTMLInputElement>) => {
     setCurrentTab(e.currentTarget.id);
